requirements.txt
django >= 4.0
djangorestframework
django-cors-headers
pyyaml
requests
djangorestframework-simplejwt